+++ src/IRremote.h
@@ -11,10 +11,10 @@
 
 #warning It seems, that you are using an old version 2.0 code / example.
 #warning This version is no longer supported!
-#warning Upgrade instructions can be found here: https://github.com/Arduino-IRremote/Arduino-IRremote?tab=readme-ov-file#converting-your-2x-program-to-the-4x-version
+#warning Upgrade instructions can be found here: "https://github.com/Arduino-IRremote/Arduino-IRremote?tab=readme-ov-file#converting-your-2x-program-to-the-4x-version"
 #warning Please use one of the new code examples from the library, available at "File > Examples > Examples from Custom Libraries / IRremote".
 #warning Start with the SimpleReceiver or SimpleSender example.
-#warning The examples are documented here: https://github.com/Arduino-IRremote/Arduino-IRremote?tab=readme-ov-file#examples-for-this-library
+#warning The examples are documented here: "https://github.com/Arduino-IRremote/Arduino-IRremote?tab=readme-ov-file#examples-for-this-library"
 #warning Or just downgrade your library to version 2.6.0.
 
 /**********************************************************************************************************************