+++ src/TinyIRReceiver.hpp
@@ -28,7 +28,7 @@
  ************************************************************************************
  * MIT License
  *
- * Copyright (c) 2022-2024 Armin Joachimsmeyer
+ * Copyright (c) 2022-2025 Armin Joachimsmeyer
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -94,10 +94,8 @@
 #endif
 
 #if defined(TRACE)
-#define LOCAL_TRACE
 #define LOCAL_TRACE_STATE_MACHINE
 #else
-//#define LOCAL_TRACE // This enables trace output only for this file
 //#define LOCAL_TRACE_STATE_MACHINE  // to see the state of the ISR (Interrupt Service Routine) state machine
 #endif
 
@@ -153,9 +151,7 @@ volatile TinyIRReceiverCallbackDataStruct TinyIRReceiverData;
  */
 extern void handleReceivedTinyIRData();
 
-#if defined(LOCAL_DEBUG)
-uint32_t sMicrosOfGap; // The length of the gap before the start bit
-#endif
+uint32_t sMicrosOfGap; // The length of the gap before the start bit, used for trace
 /**
  * The ISR (Interrupt Service Routine) of TinyIRRreceiver.
  * It handles the NEC protocol decoding and calls the user callback function on complete.
@@ -211,7 +207,7 @@ void IRPinChangeInterruptHandler(void) {
             // We are at the beginning of the header mark, check timing at the next transition
             tState = IR_RECEIVER_STATE_WAITING_FOR_START_SPACE;
             TinyIRReceiverControl.Flags = IRDATA_FLAGS_EMPTY; // If we do it here, it saves 4 bytes
-#if defined(LOCAL_TRACE)
+#if defined(TRACE) // Do not use LOCAL_TRACE here since sMicrosOfGap is read in a cpp file at TRACE
             sMicrosOfGap = tMicrosOfMarkOrSpace32;
 #endif
 #if defined(ENABLE_NEC2_REPEATS)